cleanup: refactor fedimintd
