+++ fedimintd/src/bin/main.rs
@@ -1,5 +1,4 @@
 use fedimint_core::fedimint_build_code_version_env;
-use fedimintd::Fedimintd;
 #[cfg(not(any(target_env = "msvc", target_os = "ios")))]
 use tikv_jemallocator::Jemalloc;
 
@@ -10,8 +9,10 @@ static GLOBAL: Jemalloc = Jemalloc;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    Fedimintd::new(fedimint_build_code_version_env!(), None)?
-        .with_default_modules()?
-        .run()
-        .await
+    fedimintd::run(
+        fedimintd::default_modules,
+        fedimint_build_code_version_env!(),
+        None,
+    )
+    .await
 }
