+++ fedimintd/src/envs.rs
@@ -1,54 +1,31 @@
-// Env variable to TODO
 pub const FM_DISABLE_META_MODULE_ENV: &str = "FM_DISABLE_META_MODULE";
 
-// Env variable to TODO
 pub const FM_DATA_DIR_ENV: &str = "FM_DATA_DIR";
 
-// Env variable to TODO
 pub const FM_BIND_TOKIO_CONSOLE_ENV: &str = "FM_BIND_TOKIO_CONSOLE";
 
-// Env variable to TODO
 pub const FM_BIND_P2P_ENV: &str = "FM_BIND_P2P";
 
-// Env variable to TODO
 pub const FM_P2P_URL_ENV: &str = "FM_P2P_URL";
 
-// Env variable to TODO
 pub const FM_BIND_API_ENV: &str = "FM_BIND_API";
 
-// Env variable to TODO
-pub const FM_BIND_API_WS_ENV: &str = "FM_BIND_API_WS";
-
-// Env variable to TODO
-pub const FM_BIND_API_IROH_ENV: &str = "FM_BIND_API_IROH";
-
-// Env variable to TODO
 pub const FM_API_URL_ENV: &str = "FM_API_URL";
 
-// Env variable to bind the web UI server
 pub const FM_BIND_UI_ENV: &str = "FM_BIND_UI";
 
-// Env variable to TODO
 pub const FM_BITCOIN_NETWORK_ENV: &str = "FM_BITCOIN_NETWORK";
 
-// Env variable to TODO
 pub const FM_BIND_METRCIS_ENV: &str = "FM_BIND_METRICS";
 
-// Env variable to TODO
 pub const FM_PORT_ESPLORA_ENV: &str = "FM_PORT_ESPLORA";
 
-// Can be used to absolutely override the values stored in the db
 pub const FM_FORCE_API_SECRETS_ENV: &str = "FM_FORCE_API_SECRETS";
 
-// Env variable for the bitcoind URL
 pub const FM_BITCOIND_URL_ENV: &str = "FM_BITCOIND_URL";
 
-// Env variable for the Esplora URL
 pub const FM_ESPLORA_URL_ENV: &str = "FM_ESPLORA_URL";
 
-/// Use iroh for networking
 pub const FM_ENABLE_IROH_ENV: &str = "FM_ENABLE_IROH";
 
-/// Environment variable for the session count determining when to cleanup old
-/// checkpoints.
 pub const FM_DB_CHECKPOINT_RETENTION_ENV: &str = "FM_DB_CHECKPOINT_RETENTION";
